Refactor: propogate up forbidden HTTP status on error
