+++ backend/src/main/kotlin/org/conservationco/asanahire/service/UserService.kt
@@ -3,7 +3,7 @@ package org.conservationco.asanahire.service
 import org.conservationco.asanahire.model.user.User
 import org.conservationco.asanahire.model.user.valueOfIgnoreCase
 import org.conservationco.asanahire.repository.UserRepository
-import org.conservationco.asanahire.util.handleSaveError
+import org.conservationco.asanahire.exception.handleSaveError
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.context.ReactiveSecurityContextHolder
 import org.springframework.security.core.context.SecurityContext
